Theme linter plugin refactor.  Renamed things and broke out plugins into separate classes.
